+++ src/middlewares/server.middleware.ts
@@ -1,6 +1,6 @@
 import { Request, Response, NextFunction } from 'express';
 
-console.log('import app.middleware');
+console.log('import server.middleware');
 
 export const ServerMiddleware = (req: Request, res: Response, next: NextFunction) => {
     console.log('middleware of application');